requirements.txt
annotated-types
anyio
arabic-reshaper
arxiv
asn1crypto
beautifulsoup4
blinker
Brotli
certifi
cffi
charset-normalizer
click
contourpy
cryptography
cssselect2
cycler
defusedxml
distro
duckdb
duckduckgo_search==5.3.1b1
et_xmlfile
fastapi
Flask
fonttools
fpdf
fpdf2
greenlet
h11
html5lib
huggingface_hub
idna
ipykernel
itsdangerous
Jinja2
kiwisolver
langchain
langchain-chroma
langchain-community
langchain-groq
langchain-huggingface
langchain-openai
langchain-text-splitters
lxml
Markdown
MarkupSafe
matplotlib
mysql-connector-python
narwhals
numexpr
numpy
openai
openpyxl
oscrypto
packaging
pandas
pdf2image
pillow
playwright
plotly
pycparser
pydantic
pydantic_core
pydyf
pyee
pyHanko
pyhanko-certvalidator
pyparsing
pypdf
pyphen
python-bidi
python-dateutil
python-dotenv
pytube
pytz
PyYAML
reportlab
requests
seaborn
sentence_transformers
six
sniffio
soupsieve
SQLAlchemy
ssh-import-id
starlette
streamlit
supervisor
svglib
tabulate
tinycss2
tinyhtml5
tqdm
typing-inspection
typing_extensions
tzdata
tzlocal
unstructured
uritools
urllib3
uvicorn
validators==0.28.1
weasyprint
webencodings
Werkzeug
wikipedia
xhtml2pdf
youtube_transcript_api
zopfli
